Add strategy execution logging tests for ML and enhanced strategies
